feat: implement RaffleFactory and refactor Instance architecture (#33)
